Respect XDG_CONFIG_HOME settings
If a user has XDG_CONFIG_HOME set, we should load the CircleCI
configuration from that directory, instead of from $HOME/.circlerc.
Update the docs to make this more clear. Update the error handling and
error messages to be a little more robust - for example, user.Current
often fails with cross compilation.

We also previously failed to close `f`, which is not great.