Implement --rebase=X
Frequently I want to wait to merge a change into master, but the tests
become invalid if someone else merges a change. Check whether someone
else is attempting to merge something, and if so, rebase/force push to
the branch in question. Also perform this check before marking a test
as passing or failing and then stopping the build.