Update all queries to use Circle v1.1 API
Some of the API's have changed, since we must now tell Circle whether
the project is a Bitbucket or a Github project. We could probably
refactor this further, but as far as I know, I am the only consumer of
these API's, so I don't feel too bad about changing them.

We're currently creating a rest.Client every time we do a query which
is not great but the slow part is the API querying so I'm not too
worried about it.

Implement "circle cancel" to cancel a build - we had the API for this
but not the subcommand in the binary, so add it.

Replace skratchdot with pkg/browser since I trust the maintainer of
the other one better.